feature(#52): Rename sMinus to omegaMinus and add interface to use magic parameter
add getter and setter for omegaMinus and for the magic parameter. When the magic parameter is setted, omegaMinus is changed accordingly. Every time that setOmega is called,by default omegaMinus is recomputed to keep costant the magic parameter. This behaviour can be modified putting keep_magic_constant_when_setting_omega=false